make refactoring

+++ area_triangle.rb
@@ -1,6 +1,6 @@
-puts "What is the base of your triangle"
-base = gets.chomp.to_i
-puts "What is the height of your triangle"
-height = gets.chomp.to_i
-area = base*height/2
+puts 'What is the base of your triangle'
+base = gets.chomp.to_f
+puts 'What is the height of your triangle'
+height = gets.chomp.to_f
+area = base * height / 2
 puts "Area of your triangle is #{area}"
