+++ quadratic_equation.rb
@@ -1,21 +1,21 @@
 puts "Enter a:"
-a = gets.chomp.to_i
+a = gets.chomp.to_f
 puts "Enter b:"
-b = gets.chomp.to_i
+b = gets.chomp.to_f
 puts "Enter c:"
-c = gets.chomp.to_i
+c = gets.chomp.to_f
 
-d = b**2 - 4*a*c
+d = b**2 - 4 * a * c
 
 if d >= 0
-x1 = (-b + Math.sqrt(d))/(2*a)
-x2 = (-b - Math.sqrt(d))/(2*a)
+  x1 = (-b + Math.sqrt(d)) / (2 * a)
+  x2 = (-b - Math.sqrt(d)) / (2 * a)
 end
 
 if d < 0
-	 puts "Discriminant = #{d}. There are no real roots."
+  puts "Discriminant = #{d}. There are no real roots."
 elsif d == 0 
-	puts "Discriminant = #{d}. There is one real root: #{x1}."
+  puts "Discriminant = #{d}. There is one real root: #{x1}."
 else 
-    puts "Discriminant = #{d}. There are two real roots: #{x1}, #{x2}."
+  puts "Discriminant = #{d}. There are two real roots: #{x1}, #{x2}."
 end
